requirements.txt
Django==4.2.11
django-environ
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
langchain-community
langchain
huggingface-hub
gdown
langchain-huggingface
whitenoise
colbert-ai